Progress Check
Progress check
Alexandru is a student in the 9th grade. He just learned in math class about arithmetic progressions and wants to use his programming skills to automate his homework. In the homework problem, he is given two consecutive numbers in arithmetic progression, and he must find the next three terms of the progression.

3 integers a, b and c are consecutive terms of an arithmetic progression if the difference between a and b is equal to the difference between b and c.

Input data
The numbers a and b are read from the keyboard, representing the 2 consecutive numbers in arithmetic progression.

Output data
The next 3 terms of the respective arithmetic progression will be displayed on the screen.

Restrictions and clarifications
It will be assumed that a and b are integers.

-1,000 < a, b < 1,000

Example
Input data Output data
5 2 -1 -4 -7
15 21 27 33 39