Percentage of users problem
Percentage of Users
The team wants to know the status of its platform in terms of the number of users. Knowing the number of users from the previous week and the current week, they want to know the percentage by which the number of users of the platform has increased.

They have already written a program that solves this problem. Can you write one too?

Input data
The numbers x and y are read from the keyboard, representing the number of users from the previous week and the number of users from the current week, respectively.

Output data
The program will display on the screen the percentage by which the number of platform users has increased.

Restrictions and clarifications
0 < x ≤ y

0 < y < 1000

The result is guaranteed to always be an integer with no decimals.

Example
Input data Output data
70 77 10
50 100 100