DISTANCE
Requirement
The real number axis contains all negative and positive numbers, arranged in ascending order from left to right. You are given an integer value x. Write a program to display the distance between the x coordinate point and the origin (coordinate point 0).
This distance between a given number and 0 is called the modulus.

This notion is very useful when we want to calculate distances between various points and is used a lot in video games.

Input data
A single value, x, is read.

Output data
The distance between the x-coordinate point and the 0-coordinate point will be displayed.

limitation
-100 ≤ x ≤ 100
Example
Input data Output data
45 45
-30 30