progress check
Requirement
Given four integers a, b, c and d. Check if the numbers are in a geometric progression.

Explanations
The ratio of two numbers a and b is the value of dividing a by b.

A geometric progression is a sequence of numbers with the property that the ratio of any two consecutive terms is constant (that is, the value is always the same).

This constant value is called ration.

The fact that we know that a series of numbers is in a geometric progression is useful because we can very easily find out elements such as the ration, the sum of the terms and the general term, which simplifies the solving of the problems that are based on this theme.

Example from everyday life
An everyday example of geometric progression is the amount of money in your bank account each month. For example, if you had 1% interest every month and initially 100 lei in your account, the first month you would have 100 lei in your account, in the second you would have 101, in the third 102.01 and so on...

Knowing that the monthly interest is 1% (so having a geometric progression), we can very easily calculate how much money we will have in the account after 10 years (if we did not touch it).

See here for other examples of applications of geometric progressions.

Input data
The integers a, b, c and d are read from the keyboard.

Output data
The program will output YES if the four numbers are in a geometric progression and NO if the four numbers do not meet the condition.

Restrictions and clarifications
-200 < a, b, c, d < 1,000
It is guaranteed that the ratio between a number and the one before it (if this number exists) is a non-zero integer (that is, different from 0).
Input data Output data
2 4 -8 16 NO
3 9 27 81 YES
1 5 25 125 YES