Stairs
Gigel, Cornel and Tibi are 3 friends who make a mountain route to Vârful Toaca. To get to the top he has to climb a very long ladder. Being tired, they decide to take a break. How many free steps are there between the three when they stop?

Input data
3 integer values ​​are read from the keyboard, a, b and c, corresponding to the steps on which the 3 friends stop.

Output data
The program will display on the screen the number X which represents the number of free steps between the 3 friends.

Restrictions and clarifications
0 < a, b, c < 1000
the numbers are given in some order, they are not necessarily part of an ascending or descending sequence
Examples
Input data Output data
10 7 12 3