Ascending order
Requirement
First grade student, Mihai has just learned to order numbers in ascending order, and now he wants to do as many exercises as possible. He heard that it is possible to write a program that sorts 4 natural numbers in ascending order. If they had such a program, it could be checked very easily.

Mihai is asking you to create such a program for him.

Input data
The program reads the 4 distinct numbers.

Output data
The program displays the 4 numbers in ascending order.

limitation
The numbers are between 0 and 100.
Example
Input data                      Output data
6 4 1 20                             1 4 6 20