Ascending 3 - Part 2
Dragoș's father, Marian, is a mathematics teacher at a renowned high school. So, although Dragoș is only in first grade, he is far above the level of his peers and already knows what "module of a number" means.

Since Dragoș recently learned at school how to order breeder, Marian wants to combine the information assimilated by his child. He gave Dragoș the task of ordering 3 numbers in ascending order, according to their modulus, in order to practice these two notions.

Wanting to prove to his father that he is a smart kid, he asked you to do this, because he can't handle it.

To show off your programming skills, you want to give Dragoș not only the answer, but also a program to do it automatically for him.

In mathematics, the modulus or absolute value of a number x, denoted |x|, is the unsigned number. Example: |-3| = 3.

Input data
3 whole numbers are read from the keyboard.

Output data
The program will display the numbers on the screen in ascending order according to their modulus, separated by spaces.

limitation
The numbers fall into the int data type.
The values ​​of the given numbers will be distinct in the mod.
Example
Input data                     Output data
-3 2 1                                1 2 -3