TRIANGLE
Given a number n. Display the following triangle of numbers:

n n-1 n-2 ... 2 1 (1)
n-1 n-2 ... 2 1 (2)
....
2 1 (n-1)
1
More precisely, on the first line all the numbers from n to 1 will be displayed followed by the index of the line they are on as in the example above, on the second line all the numbers from n-1 to 1 etc. The numbers on each line will be separated by spaces.

Input data
The number n is read on the keyboard.

Output data
The program will display the triangle requested above on the screen.

limitation
0 < n < 50

Example
Input data
4
Output data
4 3 2 1 (1)
3 2 1 (2)
2 1 (3)
1 (4)