Discount codes!
Story
Ionel created a program for his store that generates discount codes to attract as many customers as possible. However, for that discount code to work, Ionel has to add one more digit to the beginning.

Now Ionel needs a program that can do this for a single discount code. His friend, Cosmin, will make sure that the program you write will later be applied to all codes.

Input data
The program reads a number n and a digit c from the keyboard.

Output data
The program must display the resulting discount code after adding the digit c to its beginning, as well as the double of the new code obtained.

Restrictions and clarifications
1 ≤ n ≤ 10 000 000
1 ≤ c ≤ 9
Example
Input data
2345 1
4892 5
Output data
12345 24690
54892 109784