adjacent Even Numbers
Tudor is passionate about statistics. He makes different statistics, which are more and more strange and unprecedented. This time, he came up with the idea of ​​calculating the number of even-adjacent pairs of digits from the registration numbers of his friends in a game they used to play, to study the frequency of the obtained values.

He asked you to create a program that would make his work more efficient.

Input data
A natural number x is read from the keyboard, representing the registration number of one of Tudor's friends.

Output data
The program displays a number p on the screen, representing the number of even-adjacent pairs of digits in the number x.

Restrictions and clarifications
0 ≤ x ≤ 2 147 483 647
Example
Input data
24235792
Output data
1 5