On other planets
Requirement
It is considered a coding of a solar system consisting of several planets by a number: the first digit represents the characteristic digit of the first planet, the second digit characterizes the second planet, etc. An inhabitant of a planet can travel to another planet if the characteristic digit of the planet from which he comes is a divisor of the digit of the planet he wants to reach.

You are on the first planet of a solar system consisting of X planets and you want to travel to another planet. How many of the remaining X-1 planets can be your next destination?(X is the number of digits of n).

Input data
A single natural number, n, is read, representing the encoding of the solar system.

Output data
The answer to the question in the requirement, the number of planets the inhabitant of planet 1 can move to, will be displayed.

limitation
0 ≤ n ≤ 10 000 000
If n has at least two digits, the first digit is different from 0
Example
Input data
22354
Output data
2