Fibonacci
Given a number n. Find the nth term of the Fibonacci sequence.

Fibonacci:
was a mathematician considered as "the most talented mathematician in the West of the Middle Ages".
introduced arithmetic into the European trading system
introduced the idea of ​​operations with fractional numbers, the same denominator procedure, mixture problems, irrational numbers and recurrence relations.
approximated the value of π to 864/275.
Input data
A natural number n is read from the keyboard.

Output data
The nth term of the Fibonacci sequence will be displayed on the screen.

Remarks
In the Fibonacci number sequence, each number represents the sum of the previous two numbers, except for the first 2 values, which are 1. Thus, the sequence starts with 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, etc.
Restrictions and clarifications
1 ≤ n ≤ 40
Examples
Input data
1 1
Output data

7 13