Removing Occurrences
Requirement
Andrei, an amateur soccer player, has a collection of balls at home. He assigned a number to each ball. One day he placed all the balls in a line and realized that the numbers on the balls formed a natural number. He also noticed that all the balls with the number c written on them had a garish hue, so he decided to remove them from the row. What is the new number formed after removing all the balls that correspond to the number c? Now it's up to you to figure it out.

Input data
The number n and the number c are read from the same line, representing the number formed with the numbers of the balls, respectively the number that appears on the balls to be eliminated.

Output data
The number obtained after performing the eliminations is displayed.

limitation
n has at least 1 digit and at most 8.
If, after removing all occurrences of the digit c, no digits remain, -1 will be displayed, and if only null digits remain, 0 will be displayed.
If, following eliminations, the new number starts with one or more 0s and there are other non-zero digits in the new number, those leading 0s will also be removed.
Example
Input data
1441 4
Output data
11