Wellcode figure
Requirement
Given a natural number n. Display its Wellcode digit. The Wellcode digit is calculated by doing the product of the digits of the number, then the product of the digits of the product, and so on until the product obtained is a digit.

Input data
A single number, n, is read, having the meaning in the statement.

Output data
A single digit will be displayed, representing the Wellcode digit.

limitation
0 ≤ n ≤ 1 000 000 000
Examples
Input data
1234
Output data
8
Explanation
1 * 2 * 3 * 4 = 24
2 * 4 = 8