New friends
Requirement
Numbers have become very sociable. A number a can be friends with a number b only if b is divisible by a.

The program you write will have to read natural numbers from the keyboard until the read value is 0. How many of the read numbers can the first read number be friends with?

Input data
The string of numbers is read until it meets 0. The numbers will be separated by spaces.

Output data
The answer to the question in the requirement, the number of elements with which the first number of the string is friends, will be displayed.

limitation
A number can befriend itself
A maximum of 100,000 numbers are read until 0 is encountered
0 is not part of the string
The elements of the string are natural numbers between 0 and 1000
Example
Input data
4 8 3 5 4 0
Output data
3