Monotony to 0
Paul is playing a new computer game. In it, he must cross a mountainous area as quickly as possible. Looking at the game map, he can see all the terrain heights around him. Thus, he can choose the sequence of steps to take from where he is to his destination.

Being a mountainous area, it is possible that the road chosen by Paul goes up and then down the mountain. He thinks that it would be better to avoid such a path because it would mean more effort on his part. Likewise, after going down a slope, Paul would rather not have to go up again.

Requirement
Paul drew on the map a road he could take to reach his destination. He wrote down the height of each piece of land on a piece of paper, in the order in which he had to traverse them.

You are given the string of heights as a space-separated string of non-zero integers. The string of numbers is terminated by the value 0, which is not part of the heights (it just marks the end of the string). Specify whether the series of heights is increasing, decreasing or non-monotonic. We say that a sequence is non-monotonic if it is neither increasing nor decreasing in its entirety.

Input data
A string of natural numbers followed by the value 0 is given.

Output data
One of the messages ascending, descending or non-monotonic will be displayed.

Restrictions and clarifications
Heights are numbers between 0 and 10,000
A maximum of 1,000 numbers will be read
0 is not part of the height string, it just marks the end of the string
For simplicity, if all numbers are equal or there is no value before 0, the string is considered non-monotonic
Examples
Input data
1 2 5 5 10 11 0
16 7 3 0
1 2 2 1 0
Output data
Increasing
Descending
nonMonotone