Mountain
Requirement
George is a fan of mountain hiking. He believes that a trip to the mountain gives him a real experience if it has a single peak and more, measuring the altitude kilometer by kilometer and building a string with these values, that string is a mountain string.

A mountain string is a string in which there is a single index i ("index" = position of an element in the string) that checks all of the following properties:

2 ≤ i ≤ n - 1, where n represents the number of elements in the string
starting with the first number in the row up to and including the ith number, the numbers are in strictly ascending order.
starting with the ith number in the row up to and including the last number, the numbers are in strictly descending order.
Given a string of n numbers. To check whether the given string represents a mountain string or not.

Input data
On the first line is n, the size of the string. On the next line are n natural numbers, the elements of the string representing George's measurements.

Output data
A single number will be displayed: 1 if the given string is a mountain string, or 0 otherwise.

limitation
3 ≤ n ≤ 500
The elements of the string are natural numbers less than 1,000,000
Example
Input data
7
1 4 6 7 8 5 2
7
1 4 6 10 8 9 2
Output data
1
0