Sum of two large numbers
Requirement
Defines the adda function, which takes as parameters 3 integer strings, a, b, and the result. At the end of execution, the resulting string will contain the result of the sum of the numbers saved in a and b. All 3 strings will follow the format described above.

Function signature
The name of the function will be add
The function will have 3 integer parameters: a[], b[] and result[]
The function will be of type void
limitation
1 ≤ a[0], b[0] ≤ 1000
The elements of the two strings (except the first element) are natural numbers in the interval [0, 9] (digits)
It is guaranteed that a[a[0]] and b[b[0]] (the first digits of the 2 numbers) will not have the value 0
Example
a[] = {5, 1, 2, 2, 5, 8}; b[] = {6, 2, 2, 2, 6, 7, 8}
After calling add(a, b, result), the result[] string becomes:

result[] = {6, 3, 4, 4, 1, 6, 9}