adding a large num to another large num
Requirement
Defines the adda function, which receives as parameters 2 strings of integers, a and b. At the end of the execution, the string a will contain the result of the sum of the numbers initially saved in a and b. Both strings will follow the format described above, when the function it is also called after the call ends.

Try to solve the problem without auxiliary number strings.

Function signature
The name of the function will be add
The function will have 2 integer parameters: a[] and b[]
The function will be of type void
limitation
1 ≤ a[0], b[0] ≤ 1000
The elements of the two strings (except the first element) are natural numbers in the interval [0, 9] (digits)
It is guaranteed that a[a[0]] and b[b[0]] (the first digits of the 2 numbers) will not have the value 0
Example
a[] = {5, 1, 2, 2, 5, 8}; b[] = {6, 2, 2, 2, 6, 7, 8}
Following the call to add(a, b), the string a[] becomes:

a[] = {6, 3, 4, 4, 1, 6, 9}