Create Multiplication
Gigel learned that multiplication is actually repeated addition. He has not yet grasped this concept, and in order to understand, he needs to write down each step of the multiplication.

Requirement
Given two natural numbers a and b, write each partial result of multiplying the numbers a and b, i.e. a, 2 * a, 3 * a, ... ,b * a.

Input data
Two numbers, a and b, will be read from the keyboard, with the meanings from the statement (multiples of the first number are always written).

Output data
On the screen will be written b numbers, each partial result of the multiplication of the two numbers.

limitation
0 < a < 1,001

0 < b < 1,001

Example
Input data Output data
5 3 5 10 15
Explanation
5 * 1 = 5

5 * 2 = 10

5 * 3 = 15