Create ABC Divide
Requirement
Robin dreams of becoming a researcher. This time he asks himself a new question: "Considering the numbers n, a, b and c, is there any mathematical connection between the first n numbers that are not divisible by a, b or c?". But all he needs from you is to find out the n numbers.

Thus, given 4 numbers: n, a, b and c, display the first n numbers that are not divisible by any of the numbers a, b or c.

Input data
The 4 natural numbers, n, a, b and c are read in this order.

Output data
The n numbers that meet the given condition are displayed, on a single line, separated by a space.

limitation
1 ≤ n ≤ 1,000
2 ≤ a, b, c ≤ 1,000
The numbers displayed are guaranteed to be less than 10,000
Example
Input data Output data
6 3 4 7 1 2 5 10 11 13