Create Multiple Range
On the occasion of the city days, the mayor thought of offering a small gift to the citizens. Since he does not have enough funds to cover the expenses of the gifts for all the citizens of the city, he thought of proceeding as follows:

On each street there are houses numbered from a to b inclusive. To please at least some of the citizens, the mayor will give gifts to residents whose house number is a multiple of a chosen number, k.

Being a medalist at ONI, the mayor of the city asked you to make a program to help him with how to distribute gifts on every street.

Input data
The program will read from the keyboard the natural number k, representing the number whose multiples we are looking for, the natural number a, representing the number of the first house on a given street, and the natural number b, representing the number of the last house on the same street.

Output data
The program will display on the screen the numbers of all the houses whose residents will receive a gift from the mayor. Numbers will be separated by spaces.

Restrictions and clarifications
0 ≤ a, b < 1000
a < b
0 < k < 100
Example
Input data Output data
3 5 24 6 9 12 15 18 21 24