Create K Multiples
Andrei is a stamp collector. Since he has too many stamps and not enough storage space, he thought of giving some of them to a friend of his, who is just as passionate. Since he is attached to each individual stamp, Andrei cannot choose which stamps to give as gifts, so he thought of a way to automatically choose the stamps.

It has n number of stamps, and each stamp has an order number from 0 to n - 1. He thought of giving his friend k stamps, namely those with the highest order number less than n, which are multiples of a chosen number p.

Input data
The program reads from the keyboard the natural number n, representing the number of stamps in the collection, a number k, representing the number of stamps Andrei wants to offer, followed by another number p, the number whose multiples we are looking for.

Output data
The program will display on the screen the serial numbers of the stamps that Andrei will give to his friend, in descending order.

Restrictions and clarifications
0 < n ≤ 2500
0 < k < 25
0 < p ≤ 100
It is guaranteed that there is a solution for every test
Example
Input data Output data
158 4 11 154 143 132 121