Create Picture
Paul is passionate about collecting banknotes. He even made a folder where he collects banknotes of different countries. He wants to make a custom cover for his folder that has a square design with a border of * characters and $ characters inside, but he doesn't know how to do it automatically.

Paul asks you to write a program to generate the pattern on the cover of the folder, to then print it to the printer.

Input data
The number n is read from the keyboard, representing the length of the side of the square.

Output data
A square will be displayed on the screen with the edges formed by the * character and the interior by the $ character.

Restrictions and clarifications
0 < n < 100
Example
Input data Output data
4	****
*$$*
*$$*
****
5	*****
*$$$*
*$$$*
*$$$*
*****