Create *-diamond
Requirement
Andreea is in charge of arranging a room. This time he has to paint some diamond-shaped patterns on the walls, but he doesn't think he'll be able to do it unless he also receives the digitized sketch. Thus, she gives you the diagonal of the rhombus, and you will have to represent it.

Your task is to form a rhombus (according to the model in the example) with both diagonals of length n, the sides formed by *, and the interior containing only spaces.

Input data
A single natural number n is read, the length of the diagonals of the rhombus.

Output data
They will be displayed in lines, representing the requested "drawing".

limitation
3 ≤ n ≤ 100
It is guaranteed that n is odd
Example
For n = 5 it will show
  *
 * *
* *
 * *
  *