Create Prime Factors
Mihnea is a mathematics teacher. He is going to teach his students the decomposition into prime factors. Because he is passionate about his subject, but also about computer science, he creates a program to help his students understand the subject more easily. The program decomposes any number n into prime factors.

Try to make a program similar to the one made by Mihnea!

Input data
The program reads a natural number n from the keyboard.

Output data
The program displays on the screen the decomposition into prime factors of the number n in the form x^p. On each line, a prime factor will be displayed at the power at which it decays.

Restrictions and clarifications
1 < n ≤ 1,000
x will always be prime and p is the power to which x is in the prime factorization.
in the decomposition into prime factors, the factors will be displayed in ascending order, each on different lines.
Example
Input data Output data
44 2^2
11^1