Even and odd positions
Requirement
Andrei learned today what an even or odd number is. To make sure he understood, his teacher gave him a string of numbers and asked him to find the difference between the sum of the even numbers and the sum of the odd numbers.

Input data
On the first line is n, the number of elements in the string. On the next line are n integers, the elements of the string.

Output data
A single number will be displayed, the result of the required difference.

limitation
1 ≤ n ≤ 100,000
Although he is only a small learner, Andrei already knows operations with whole numbers
The string contains integers between -1000 and 1000
Example
Input data Output data
5
-4 3 5 7 8 1