Reverse
Requirement
Dalina has n balls placed in a line, each ball having a natural number written on it. She reads the values ​​on the balls from left to right and writes them on a sheet. Being very busy at this time, Dalina asks you to help her reorganize and reverse the string of numbers before returning the sheet to her.

Input data
On the first line is n, the number of balls. On the next line are n natural numbers, representing the values ​​written by Dalina on the sheet.

Output data
n numbers will be displayed, representing the reversed string. The n numbers will be on a single line, separated by a space.

limitation
1 ≤ n ≤ 100,000
The string contains natural numbers less than 1,000,000
The requirement of the problem is to store the elements read into a string and reverse them in memory, not just display them from right to left!
This problem can arise during job interviews.
Example
Input data Output data
5
4 3 5 7 8 8 7 5 3 4