assortment
Each inhabitant of a city donated an amount of money for the creation of a center of excellence in informatics. In order to determine what is the largest value that has been donated, the mayor of the city has hired you to create a program that sorts all the donated values ​​in ascending order. Since he can't afford to waste time, he asked you to make a program that will perform the sorting in the shortest possible time.

Input data
A number n is read from the keyboard, followed by a string of n natural numbers.

Output data
The program will display the elements of the sorted string in ascending order on the screen.

Restrictions and clarifications
0 < n ≤ 100,000
the elements of the string are numbers between 0 and 500, inclusive.
Example
Input data Output data
10
15 47 98 23 145 74 89 32 1 74 1 15 23 32 47 74 74 89 98 145