Building Sorting
An architect was hired to restore all the old buildings of the city. He was handed a document with all the buildings in need of restoration, respectively a list with the height of each building. As an important event is taking place in the city soon, the architect has been asked to restore as many buildings as possible by the date of the event.

To ensure that he manages to restore as many buildings as possible in the shortest possible time, he wants to start with the shortest buildings. You offered to help him by writing a program to sort the heights of buildings in ascending order.

Input data
A natural number n is read from the keyboard, representing the total number of buildings to be restored, respectively a string of n integers, representing the heights of the buildings - in millimeters.

Output data
The program will display on the screen the heights of the buildings, sorted in ascending order, separated by spaces.

Restrictions and clarifications
1 ≤ n ≤ 1,000
building heights are integers in the range [-2 000 000 000, 2 000 000 000]
Example
Input data Output data
7
41730 939963 571901 9571 553691 751366 367571 9571 41730 367571 553691 571901 751366 939963