Stack
Requirement
Given a string of n integers on which the following operations are performed: removing the last element from the string (pop) and adding an element x to the end of the string (push x). To display the string after applying the given operations.

The data structure introduced above, on which only push x and pop operations can be performed, is called a stack. The stack is a data structure often used in computer science and is said to be last in, first out because of the way we access data stored in it. It is used both to implement recursion in programming languages ​​and as an auxiliary structure for traversing more complicated data structures such as trees and graphs.

Input data
It reads the natural number n, representing the number of elements of the original string, and n integers, representing the elements of the string. Then m values ​​of k are read where k can have the value 1 or 2. If k has the value 1, an integer x will be read immediately after it and the push x operation will be performed. If k has the value 2, the pop operation will be performed.

Output data
The string of numbers obtained after applying the given operations to the original string will be displayed on the screen. The first line will display the number t, representing the number of elements of the new string, and the second line will display the t elements of the string, separated by spaces.

Restrictions and clarifications
n and m are natural numbers in the range [1, 10 000].
Each element in the string is in the range [-2 000 000 000, +2 000 000 000].
k will always have the value 1 or 2.
x has values ​​in the range [-2 000 000 000, +2 000 000 000].
There will be no type 2 operations when the stack is empty.
Example
Input data Output data
7
5 -4 0 -7 7 7 2
5
2
1 100
2
1 0
1 3 8
5 -4 0 -7 7 7 0 3