queue
Requirement
Everyone knows that office work can become very stressful and tiring, especially at the end of the day. Being behind schedule, Andrei hopes that he will be able to finish the job as soon as possible so that he can go home.

He has a stack of sheets on the table that he has to sign, and each sheet has an identification number, a natural number. From time to time, Mihai also appears in the office, who also stays behind schedule and composes documents. Mihai adds one sheet below the existing ones on the table, without changing their order. Andrei takes a sheet from the top of the stack, signs it and puts it aside.

To help him get the job done faster, you offered to create a robot for him to manage the stack of sheets. This means that Andrei will only have to sign the papers that are placed under his nose, and the robot will take care of adding documents received from Mihai to the stack and extracting the document that he will give to Andrei to sign.

Of course, since the day is short, Andrei will not be able to finish all the work even with the help of your robot. The robot recorded m operations it did during the hours it was active. Displays the ID numbers of the remaining sheets after the m operations.

Input data
On the first line are two numbers, n and m.

On the following lines are n natural numbers, the identification numbers of the sheets initially on the table (the first number is associated with the first sheet, the second number is associated with the second sheet...).

On the following m lines are the operations to be done:

1 if the first sheet is to be extracted
2 followed by a value x, if a new sheet is to be inserted at the end, this one having the identification number x.
Output data
The first line will show the number of sheets left on Andrei's desk after the execution of the m operations, and on the second line, their identification numbers.

limitation
1 ≤ n ≤ 900,000
1 ≤ m ≤ 400,000
Identification numbers are between 0 and 10,000, inclusive
If the string contains no more elements and a type 1 operation is to be performed, the operation will not be performed.


Example
Input data Output data
4 3
30 21 81 26
2 68
2 64
1 5
21 81 26 68 64