Missing Element 1
Transmission of data from one computer to another within a network is done "as if by magic" in the case of high-speed connections. However, when you live on a street where no one has laid fiber optic and you're the only one who wants to watch TED Talks on YouTube, streaming starts to feel less magical.

Each piece of information is transmitted over the Internet in the form of a group of packets. If you want to stream video from YouTube, the fact that a packet is lost on the way from time to time can go unnoticed and you can watch the desired video successfully.

Suppose you could observe the packets arriving at your computer in a period of 1 minute. We can represent these packets as a string of integers. Also, the other side of the network can see the packets that have been sent.

Requirement
You are given a string of n integers, representing the packets that were sent, and another string of n - 1 integers, representing the packets that successfully reached you. The elements in the second string will be in the same order as the first, except for one element in the first string that will be missing from the second. Identify the missing element.

Challenge: Can you solve the problem by looping through the elements in each string just once? 😎

Input data
On the first line the number n will be found, and on the next two lines the elements of the two strings.

Output data
The program will display the number X on the screen, which represents the missing element in the second string.

Specifications and restrictions
1 ≤ n ≤ 1,000
the elements in the string will not have values ​​higher than 100 000 or lower than -100 000
There is a solution to the problem that uses only one forum (excluding reading). Can you find it?
Examples
Input data Output data
10
34 32 45 -67 89 21 34 5 7 9
34 32 45 89 21 34 5 7 9 -67