Missing Element 3
Curious as to why the internet at your home is getting slower every day after 6pm, you chat with your neighbors and find out they've recently subscribed to Netflix. Coincidentally, you and they both get home after 6pm every day and the network suddenly gets congested.

If in a previous problem you tried to find the packet from a string of n transmitted packets that is lost on the way and never reaches its destination, this time m packets are lost on the way.

Requirement
You are given a string of n integers, representing the packets sent, and another string of n - m integers, representing the packets that successfully reached you. The elements in the second vector will be in the same order as those in the first, except that m elements from the first array will be missing from the second. It identifies the missing elements, bearing in mind that they are not necessarily in consecutive positions in the first string.

Input data
On the first line you will find the numbers n and m, separated by a space, and on the next two lines the elements of the two strings.

Output data
The program will display on the screen, in the order in which they appear in the first line, separated by a space, the m elements that are missing from the second line. If no data was lost between the two strings, the program will display the message "No data lost" on the screen.

Specifications and restrictions
1 ≤ n ≤ 100,000
0 ≤ m < n
the elements in the string will not have values ​​higher than 100 000 or lower than -100 000
Examples
Input data Output data
7 3
12 32 45 87 100 29 34
12 87 100 34 32 45 29