Missing Element 2
Curious to find out more about the reasons why your internet speed is not constant, you started reading about different network data transmission protocols. Among other things, you learned that the UDP protocol does not guarantee that packets will reach their destination entirely. This explains the problem you sometimes have when you talk to your mother on Skype and some sentences don't reach you anymore.

However, you found out another interesting thing: UDP also does not guarantee that packets reach their destination in the order they were sent. Is that why you sometimes feel like the discussion is going in too many directions at once?

Requirement
You are given a string of n integers, representing the packets that were sent, and another string of n - 1 integers, representing the packets that successfully reached you. The elements in the second string will NOT be in the same order as those in the first, but all the values ​​in the second string are also found in the first. Identifies the element found in the first string but missing from the second.

Input data
On the first line the number n will be found, and on the next two lines the elements of the two strings.

Output data
The program will display the number x on the screen, which represents the missing element in the second string.

Specifications and restrictions
1 ≤ n ≤ 1,000
the elements in the string will not have values ​​higher than 100 000 or lower than -100 000
Examples
Input data Output data
10
32 34 89 -67 45 21 34 5 9 7
34 32 45 89 34 21 5 7 9 -67