Missing elements 4
In the previous problem you studied the happy case where a single packet from a group of n packets transmitted over the network using the UDP protocol is lost on the way. This time, the level of network congestion causes not just one, but m packets to be lost! Can you identify the lost packs as quickly as possible, before a new season of Black mirror starts and even fewer packs reach you?

Requirement
You are given a string of n integers, representing the packets sent, and another string of n - m integers, representing the packets that successfully reached you.

The elements in the second string will NOT be in the same order as those in the first. However, if there are several packages with the same value, they arrive in the order of the positions in the original string.

Additionally, all values ​​in the second string are also found in the first. It identifies the missing elements, bearing in mind that they are not necessarily in consecutive positions in the first string.

Input data
On the first line you will find the numbers n and m, separated by a space, and on the next two lines the elements of the two strings.

Output data
The program will display on the screen, in the order in which they appear in the first line, separated by a space, the m elements that are missing from the second line. If no data was lost between the two strings, the program will display the message No Data Lost on the screen.

Specifications and restrictions
1 ≤ n ≤ 1 200
0 ≤ m < n
the elements in the string will not have values ​​higher than 100 000 or lower than -100 000
Examples
Input data Output data
7 3
12 34 45 29 100 87 32
100 87 12 34 45 29 32